rename jose.ts in thumbprint.ts
